+++ bin/handler-graphite-notify.rb
@@ -3,7 +3,7 @@
 # Released under the same terms as Sensu (the MIT license); see LICENSE
 # for details
 #
-# This will send the check status (0,1,2) to a graphite metric when a check event state changes
+# This will send a 1 to a graphite metric when an event is created and 0 when it's resolved
 # See http://imansson.wordpress.com/2012/11/26/why-sensu-is-a-monitoring-router-some-cool-handlers/
 
 require 'sensu-handler'
@@ -14,7 +14,7 @@ class Resolve < Sensu::Handler
     port = settings['graphite_notify']['port'] ? settings['graphite_notify']['port'].to_s : '2003'
     graphite = Graphite.new(host: settings['graphite_notify']['host'], port: port)
     return unless graphite
-    prop = @event['check']['status']
+    prop = @event['action'] == 'create' ? 1 : 0
     message = "#{settings['graphite_notify']['prefix']}.#{@event['client']['name'].gsub('.', '_')}.#{@event['check']['name']}"
     message += " #{prop} #{graphite.time_now + rand(100)}"
     begin